requirements.txt
torch 
torchvision 
torchaudio
tensorboard
tqdm
PyYAML
scipy